+++ src/CNBot.Core/Dtos/TGUpdateDTO.cs
@@ -8,7 +8,7 @@ namespace CNBot.Core.Dtos
    public class TGUpdateDTO
     {
         [JsonProperty("update_id")]
-        public int UpdateId { get; set; }
+        public long UpdateId { get; set; }
         [JsonProperty("message")]
         public TGMessageDTO Message { get; set; }
     }
