+++ src/CNBot.Core/Entities/Chats/ChatMember.cs
@@ -6,7 +6,7 @@ namespace CNBot.Core.Entities.Chats
     {
         public Chat Chat { get; set; }
         public long ChatId { get; set; }
-        public int TGUserId { get; set; }
+        public long TGUserId { get; set; }
         public DateTime Created { get; set; }
         public ChatMemberStatusType Status { get; set; }
         public ChatMemberPermissionType Permissions { get; set; }
