+++ src/CNBot.Core/Entities/Users/User.cs
@@ -4,7 +4,7 @@ namespace CNBot.Core.Entities.Users
 {
     public class User : BaseEntity
     {
-        public int TGUserId { get; set; }
+        public long TGUserId { get; set; }
         public string UserName { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
